algorithm/leetcode/array/sum-of-even-numbers-after-queries_test.go
package array

import (
	"reflect"
	"testing"
)

func Test_sumEvenAfterQueries(t *testing.T) {
	type args struct {
		A       []int
		queries [][]int
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "test1",
			args: args{
				A: []int{1, 2, 3, 4},
				queries: [][]int{
					{1, 0},
					{-3, 1},
					{-4, 0},
					{2, 3},
				},
			},
			want: []int{8, 6, 2, 4},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := sumEvenAfterQueries(tt.args.A, tt.args.queries); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("sumEvenAfterQueries() = %v, want %v", got, tt.want)
			}
		})
	}
}
